Added the player and the movement code
I added what is to be a place holder for the main character. I also added gravity to pull the player down. I added code for the event listeners so the player can move and jump around. I mostly had trouble getting the right keys to work and getting the player to jump and not just be stuck in the sky.